[refactor]: refactor kms tests
