+++ sbercloud/acceptance/kms/data_source_sbercloud_kms_data_key_test.go
@@ -1,7 +1,8 @@
-package sbercloud
+package kms
 
 import (
 	"fmt"
+	"github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/acceptance"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
@@ -12,8 +13,8 @@ var datakeyAlias = fmt.Sprintf("tf_key_alias_%s", acctest.RandString(5))
 
 func TestAccKmsDataKeyV1DataSource_basic(t *testing.T) {
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccKmsDataKeyV1DataSource_key,
